Function to return ACE bounds from P(X = 1 | Z), P(Y = 1 | Z), or P(X,Y | Z).

Using the polymake program, a matrix \(A\) is created such that \(A \cdot p\) will have non-negative elements if the probabilities \(p\) specify a probability distribution that does not violate the IV model. \(p\) can be probabilities of either P(X = x | Z = z) and P(Y = y | Z = z), or P(Y = y, X = x | Z = z).

get_bounds(
  gammas = NULL,
  thetas = NULL,
  zetas = NULL,
  stop = TRUE,
  warning = TRUE,
  from_polymake = NULL,
  x_mono = FALSE,
  y_mono = FALSE
)

Arguments

gammas
vector of length \(k\) giving \(P(Y = 1 | Z = z), z = 0, ..., (k - 1)\)

thetas
vector of length \(k\) giving \(P(X = 1 | Z = z), z = 0, ..., (k - 1)\)

zetas
array of dimensions \(2\cdot 2\cdot k\) giving P(Y = y, X = x | Z = z) at position (y+1,x+1,z+1),

stop
logical; should function stop with error if constraints are not met?

warning
logical; if stop is FALSE, this specifies whether a warning should be returned or not. Default: TRUE

from_polymake
result of read_polymake_results. If NULL (default), x_mono and y_mono must be specified, and previously created matrices will be used.

x_mono
logical; should P(X = 1 | Z = z) be assumed monotonically increasing? (Default: FALSE)

y_mono
logical; should P(Y = 1 | Z = z) be assumed monotonically increasing? (Default: FALSE)

Value

A list with the following elements:

lower_bounds: a vector of values that the lower bound must be greater than

upper_bounds: a vector of values that the upper bound must be smaller than

interval: a vector with the lower bound (max(lower_bounds)) and upper bound (min(upper_bounds)).

width: simply the width of the interval

constraints_violated: a logical value that indicates if any constraints are violated (TRUE) or not (FALSE)

assumptions: logical vector of length two giving x_mono and y_mono. For bookkeeping.