P(X,Y|Z) from P(X|Z), P(Y|Z), and Cov(X,Y|Z)

For specified values of \(P(X = 1 | Z = z)\), \(P(Y = 1 | Z = z)\), and \(\text{Cov}(X, Y | Z = z)\), \(z = 0,...,k-1\), the joint \(P(Y = y, X = x | Z = z)\) is reconstructed.

joint_from_marginal(
  thetas,
  gammas,
  cov,
  stop = TRUE,
  message = TRUE,
  x_mono = FALSE,
  y_mono = FALSE,
  from_polymake = NULL,
  return_bounds = FALSE
)

Arguments

thetas
vector of length n_z_levels giving P(X = 1 | Z = z), z = 0, ..., (n_z_levels - 1).

gammas
vector of length n_z_levels giving P(Y = 1 | Z = z), z = 0, ..., (n_z_levels - 1).

cov
vector of length n_z_levels giving Cov(X, Y | Z = z), z = 0,..., (n_z_levels - 1).

stop
logical; if TRUE (default), an error is thrown if the resulting probability distribution is invalid, or does not satisfy the IV inequalities. Otherwise, a warning is printed, and all probabilities returned set to NA.

message
logical; only used when stop = TRUE. If false, no warning message is printed when the resulting probability distribution is invalid. (Default: TRUE)

from_polymake
matrix created by the polymake program with the constraints a valid joint distribution should meet. If NULL (default), a previously prepared matrix is used if data seems to match one of these.

return_bounds
logical; should the resulting bounds be included in the output?

mono_x
logical; should \(P(X = 1 | Z = z)\) be monotonically increasing?

mono_y
logical; should \(P(Y = 1 | Z = z)\) be monotonically increasing?

Value

A \(2x2xk\) array with \(P(Y = y, X = x | Z = z)\) as entry (y+1, x+1, z+1).