Function to transform etas and deltas to gammas and thetas

This is a helper function that allows us to transform extreme values of \(P(Y = 1 | X = x, U)\), \(P(X = 1 | Z = z, U)\) into values of \(P(Y = y | Z = z, U)\) and \(P(X = x | Z = z, U)\), or \(P(Y = y, X = x | Z = z, U)\). It is used to programmatically generate vertices needed in the polymake program.

perform_transformation(
  etas,
  deltas,
  data_format = c("bivariate", "trivariate"),
  include_alpha = FALSE
)

Arguments

etas
vector of length 2 giving values of P(Y = 1 | X = 0, U), P(Y = 1 | X = 1, U)

deltas
vector of same length as number of categories for IV giving P(X = 1 | Z = z, U), z = 0, ..., k, k = number of IV categories minus 1.

data_format
one of 'bivariate' or 'trivariate'. If bivariate, output will give extremes in terms of P(Y|Z) and P(X|Z). If trivariate, output will give extremes in therms of zetayxz = P(Y = y, X = x | Z = z).

include_alpha
logical; whether to include column for alpha or not.