Valid range of Covs

From values of P(X = 1 | Z = z) and P(Y = 1 | Z = z) (for all z = 0, ..., n_z_levels - 1), this function returns ranges of valid values for the quantities Cov(X,Y|Z = z) and Cov(X,Y|Z = z1) - Cov(X,Y|Z = z2), such that there probability distribution defined by the input and any set of covariances chosen to fulfill the constraints will be valid, and not violated the IV inequalities.

potential_covs(thetas, gammas, x_mono = FALSE)

Arguments

thetas
vector of length n_z_levels giving P(X = 1 | Z = z), z = 0, ..., (n_z_levels - 1).

gammas
vector of length n_z_levels giving P(Y = 1 | Z = z), z = 0, ..., (n_z_levels - 1).

x_mono
logical; should monotonicity of P(X = 1 | Z = z) be assumed?

Value

A list with four elements:

thetas and gammas provided

potential_cov_range is a tibble with three columns: min and max give bounds on Cov(X,Y|Z=z) for the value of z

constraintsis a tibble with four columns: min and max give bounds on Cov(X,Y|Z = z1) - Cov(X,Y | Z = z2) for the corresponding values of z1 and z2.