Sample valid distributions P(X,Y|Z)

Returns a n x 4 tibble with one id column, and one column for each level of Z containing a random draw of a potential value of Cov(X,Y|Z=z). These values are drawn such that the joint conditional distribution P(X,Y|Z) that can be constructed from these and the values of P(X = 1 | Z) and P(Y = 1 | Z) that were provided to potential_covs satisfy the IV inequalities.

sample_joint_probs(
  pot_covs,
  n = 1000,
  max_rejections = Inf,
  x_mono = FALSE,
  y_mono = FALSE,
  from_polymake = NULL,
  return_bounds = FALSE,
  print_progress = TRUE,
  print_as_progress = NULL
)

Arguments

pot_covs
output from potential_covs()

n
number of samples to draw

max_rejections
number of rejections at which point we move on

x_mono
should it be assumed that P(X = 1 | Z = z) is monotonically increasing?

y_mono
should it be assumed that P(Y = 1 | Z = z) is monotonically increasing?

from_polymake
if NULL (default), we look through the pre-constructed matrices to find one that matches our setup. Otherwise, a matrix can be provided.

return_bounds
logical; should the bounds be returned for each sampled joint distribution? (Default: FALSE)

print_progress
logical; should process be printed?

print_as_progress
if not NULL (default), whatever is provided will be printed in beginning. Useful for batch runs.

Value

n x 4 tibble with the following columns:

id: simply ids the run, i.e. is 1:n

covs: list of vectors where each vector gives the sampled Cov(X,Y|Z = z) for each z.

joint: list of 2 x 2 x k arrays giving the joint conditional distribution reconstructed using the drawn covariances

n_rejected: number of draws that were rejected due to the reconstructed conditional distribution not satisfying the constraints.