Simulate data from IV model

This function allows the user to simulate data from an IV model with an arbitrary number of IVs with the option to include IVs that are dependent on each other, and IVs with a direct effect on the outcome.

simulate_data(
  sample_size = 10000,
  IVs_ps = list(c(0.5, 0.5), c(0.5, 0.5)),
  indIVs_on_X = NULL,
  indIVs_on_Y = rep(0, length(IVs_ps)),
  U = distributions3::Bernoulli(0.5),
  X_intercept = 0.05,
  U_on_X = 0.1,
  Y_intercept = 0.05,
  X_on_Y = 0.05,
  U_on_Y = 0.1,
  rho = NULL,
  depIVs_on_X = NULL,
  dep_probs = c(0.5, 0.5),
  depIVs_on_Y = NULL
)

Arguments

sample_size
default 10000; number of observations to simulate.

IVs_ps
list of numeric vectors. Length of list specifies number of independent IVs to include. Each vector gives probabilities of \(P(Z = i)\), and must therefore sum to 1.

indIVs_on_X
numeric vector specifying the effects of the independent IVs on X (i.e. \(\gamma_1, ..., \gamma_n\)).

indIVs_on_Y
numeric vector specifying the effects of the independent IVs on Y (i.e. \(\beta_1, ..., \beta_n\)). Default: 0

U
distributions3 object specifying the distribution to be used for the unmeasured confounder. Default is Bernoulli(0.5).

X_intercept
numeric value to use for \(\gamma_x\). Default: 0.05.

U_on_X
numeric value to use for \(\alpha_U\). Default: 0.1.

Y_intercept
numeric value to use for \(\gamma_y\). Default: 0.05.

X_on_Y
numeric value to use for \(\beta_X\). Default: 0.05.

U_on_Y
numeric value to use for \(\beta_U\). Default: 0.1.

rho
numeric value specifying \(\rho\). Default: NULL.

depIVs_on_X
numeric vector specifying the effects of the dependent IVs on X (i.e. \(\alpha_{D1}\) and \(\alpha_{D2}\)).

dep_probs
numeric vector. The length specifies the number of categories for the IVs, and the values give \(P(Z_i = z)\).

depIVs_on_Y
numeric vector specifying the effects of the dependent IVs on Y (i.e. \(\beta_{D1}\) and \(\beta_{D2}\)).

Details

Data are simulated from a logistic model as follows:

A binary confounder \(U\) is simulated with user specified \(P(U = 1)\).

A user specified number of independent IVs (\(Z_1, ..., Z_p\)) that are either di- or trichotomous. Probabilities \(P(Z_j = z), z=0,1\) (and \(P(Z_j = 2)\) if trichotomous case is considered) are specified by the user.

Optionally, a pair of dependent IVs, \(Z_{D1}, Z_{D2}\), simulated as follows:

User specifies \(\rho\)

\((N_{D1}, N_{D2}) \sim N\left(\begin{pmatrix} 0 \\ 0 \end{pmatrix},\begin{pmatrix} 1 & \rho \\ \rho & 1 \end{pmatrix} \right)\), then $$ Z_{Di} = \left\{\begin{array}{cl} 0 & \text{ if } N_{Di} \le z_{p_1} \\ 1 & \text{ if } z_{p_0} < N_{Di} \le z_{p_0 + p_1} \\ 2 & \text{ otherwise} \end{array} \right. , $$ where \(z_k\) is the \(k'th\) percentile of the standard normal (\(P(Z \le z_k) = k\) when \(Z \sim N(0,1)\)).

A binary treatment is simulated as \(X | Z_1,...,Z_p, Z_{D1}, Z_{D2}, U \sim \text{Bernoulli}(p_X)\), where $$\text{logit}(p_X) = \gamma_0 + \gamma_1 \cdot Z_1 + ... + \gamma_p \cdot Z_p + \gamma_{D1} \cdot Z_{D1} + \gamma_{D2} \cdot Z_{D2} + \gamma_U \cdot U$$, and all parameters \(\gamma_j\) are user specified.

A binary outcome is simulated as \(Y | X, U, Z_1,...,Z_p, Z_{D1}, Z_{D2}, U \sim \text{Bernoulli}(p_Y)\), where $$\text{logit}(p_Y) = \beta_0 + \beta_X \cdot X + \beta_1 \cdot Z_1 + ... + \beta_p \cdot Z_p + \beta_{D1} \cdot Z_{D1} + \beta_{D2} \cdot Z_{D2} + \beta_U \cdot U$$, and all parameters \(\beta_j\) are user specified.